Activity_Module/requirements.txt
# NeuroLens - Activity Classifier Dependencies
# Install with: pip install -r requirements.txt

# --- Web Framework ---
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0

# --- Screen Capture ---
pywinctl>=0.0.42
mss>=9.0.1
Pillow>=10.3.0

# --- ML Core ---
torch>=2.2.0
torchvision>=0.17.0

# --- Transformers & Models ---
transformers>=4.40.0
accelerate>=0.30.0          # device_map="auto" support
sentencepiece>=0.2.0        # Qwen tokenizer
einops>=0.7.0               # Florence-2 dependency
timm>=0.9.16                # Florence-2 vision backbone

# --- ONNX Export for CLIP (optional, CPU speedup) ---
# optimum[onnxruntime]>=1.19.0

# --- Testing ---
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0               # TestClient for FastAPI
